+++ src/main/java/animores/serverapi/security/BlackListToken.java
@@ -1,4 +1,4 @@
-package animores.serverapi.config.security;
+package animores.serverapi.security;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
