+++ src/main/java/animores/serverapi/security/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package animores.serverapi.config.security;
+package animores.serverapi.security;
 
 import org.springframework.data.repository.CrudRepository;
